Fixed - TreeGui was returning integers instead of doubles for real widget values and limits. Fixed - TreeGuiWidget::addDoubleSpin was calling 'addIntegerSpin' rather than 'addDoubleSpin'. Fixed - Titles of custom dialogs were not set. Fixed - Generation of dummy terms was broken, resulting in an endless loop. Added extra option for dummy term generation in GUI, allowing terms to be added to all patterns.

+++ trunk/src/parser/treegui.h
@@ -190,11 +190,11 @@ class TreeGuiWidget
 	// Return current integer value
 	int valueI();
 	// Return current double minimum
-	int minimumD();
+	double minimumD();
 	// Return current double maximum
-	int maximumD();
+	double maximumD();
 	// Return current double value
-	int valueD();
+	double valueD();
 	// Return current text
 	const char *text();
 	// Return head of items list
